# src/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Viagem365",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Token de Autenticação"
    }
  },
  "paths": {
...
    "/home/{id}": {
      "put": {
        "tags": [
          "Home"
        ],
        "description": "Desativa o status de um usuário ao realizar o logout.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/home/totalDestinos": {
      "get": {
        "tags": [
          "Home"
        ],
        "description": "Retorna a contagem total de destinos no sistema.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}